+++ test/deeplink_x_test.dart
@@ -131,6 +131,11 @@ void main() {
         expect(action, isA<App>());
       });
 
+      test('Moovit', () {
+        final action = Moovit.open();
+        expect(action, isA<App>());
+      });
+
       test('Neshan', () {
         final action = Neshan.open();
         expect(action, isA<App>());